(feat): add directory listing, refactor config api
In order to have a more consistent config api I removed the ability to pass
livereload as a number, but added the ability to check if it is an object.
You can still use the true/false shorthand for the livereload middleware.
Added the ability to enable, disable, configure the serve-index directory
listing middleware. Added some basic tests for shorthand enable/disable
syntax for livereload and directoryListing middleware.

Since I removed the livereload property as a number from the config, that
is a breaking change. This commit fixes #3.